Added documentation

+++ Image_Transformation/Converter/ToNegativeConverter.cs
@@ -4,6 +4,9 @@ using System.Windows.Data;
 
 namespace Image_Transformation
 {
+    /// <summary>
+    /// Changes an integer to negative in a view property.
+    /// </summary>
     public class ToNegativeConverter : IValueConverter
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
@@ -17,7 +20,7 @@ namespace Image_Transformation
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            if (value is double number && number < 0)
+            if (value is int number && number < 0)
             {
                 return ToggleNumber(number);
             }
